refactor: chnaged buttons path
